feat(chat): make user message separate component, pass form_value on message edit
