+++ apps/chat/src/components/Chat/ChatMessage/UserMessageContent.tsx
@@ -2,16 +2,18 @@ import { memo, useMemo } from 'react';
 
 import classNames from 'classnames';
 
+// import { ErrorMessage } from '@/src/components/Common/ErrorMessage';
+import { getFormButtonType } from '@/src/utils/app/form-schema';
 import { isSmallScreen } from '@/src/utils/app/mobile';
+
+import { FormButtonType } from '@/src/types/chat';
 
 import { useAppSelector } from '@/src/store/hooks';
 import { SettingsSelectors } from '@/src/store/settings/settings.reducers';
 import { UISelectors } from '@/src/store/ui/ui.reducers';
 
 import {
-  DialMessageFormSchema,
   DialWidgets,
-  FormSchemaButtonOption,
   FormSchemaPropertyValue,
   Message,
 } from '@epam/ai-dial-shared';
@@ -20,38 +22,35 @@ const isFormActionReply = (message: Message) => {
   return !!message.custom_content?.form_value;
 };
 
-const isPopulateButton = (option: FormSchemaButtonOption) => {
-  return !!(
-    option['dial:widgetOptions']?.populateText &&
-    !option['dial:widgetOptions']?.submit
-  );
-};
-
 const getFormActionReplyWidgets = (
   message: Message,
   index: number,
   allMessages: Message[],
 ) => {
-  if (!isFormActionReply(message)) return [];
-  const schema = allMessages[index - 1]?.custom_content
-    ?.form_schema as DialMessageFormSchema;
+  const schema = allMessages[index - 1]?.custom_content?.form_schema;
+
+  if (!isFormActionReply(message) || !schema) return [];
+
   const formValue = message.custom_content?.form_value as Record<
     string,
     FormSchemaPropertyValue
   >;
 
-  return Object.entries(schema.properties)
+  return Object.entries(formValue)
     .map(([key, value]) => {
-      const targetOption = value.oneOf?.find(
-        (option) => option.const === formValue[key],
+      const targetOption = schema?.properties?.[key].oneOf?.find(
+        (option) => option.const === value,
       );
+
+      if (!targetOption) return { isVisible: false };
 
       return {
         property: key,
-        widget: value['dial:widget'],
-        value: formValue[key],
+        widget: schema.properties[key]['dial:widget'],
+        type: getFormButtonType(targetOption),
+        value,
         label: targetOption?.title,
-        isVisible: !!targetOption && !isPopulateButton(targetOption),
+        isVisible: getFormButtonType(targetOption) !== FormButtonType.Populate,
       };
     })
     .filter(({ isVisible }) => isVisible);
@@ -77,6 +76,10 @@ const _UserMessageContent = ({
     [message, allMessages, messageIndex],
   );
 
+  const isFormReply = isFormActionReply(message);
+
+  // const isSchemaMissing = isFormReply && !allMessages[messageIndex - 1]?.custom_content?.form_schema;
+
   return (
     <div
       className={classNames('prose min-w-full flex-1 whitespace-pre-wrap', {
@@ -86,7 +89,7 @@ const _UserMessageContent = ({
       })}
     >
       <span>{message.content}</span>
-      {!!userReplyProperties.length && (
+      {isFormReply && (
         <div className="mt-2 flex items-center gap-2">
           {userReplyProperties
             .filter(({ widget }) => widget === DialWidgets.buttons)
@@ -99,6 +102,10 @@ const _UserMessageContent = ({
                 {property.label ?? property.value}
               </button>
             ))}
+
+          {/*{isSchemaMissing && (*/}
+          {/*  <ErrorMessage error="Message is missing required schema" />*/}
+          {/*)}*/}
         </div>
       )}
     </div>
